Only try to use eventEmitter if it exists (#10)
Not every context event has a related eventEmitter (for example `pluginloaderror`), which results in a TypeError when trying to call `isObserved()` on `undefined`.

The practical result of this is that a plugin failing to load—such as through a flaky network request—results in a critical error message with seemingly unrelated and hard-to-trace content.

Rather than getting a runtime error, we should just not try to emit an event if we know `eventEmitter` isn't set.

#patch